services/oxygen-calculator/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
numpy==1.26.4
python-dotenv==1.0.0
